Struct sc_authority_discovery::AuthorityDiscovery

pub struct AuthorityDiscovery<Client, Network, Block> where
    Block: BlockT + 'static,
    Network: NetworkProvider,
    Client: ProvideRuntimeApi<Block> + Send + Sync + 'static + HeaderBackend<Block>,
    <Client as ProvideRuntimeApi<Block>>::Api: AuthorityDiscoveryApi<Block>,  { /* fields omitted */ }

An AuthorityDiscovery makes a given authority discoverable and discovers other authorities.

Methods

impl<Client, Network, Block> AuthorityDiscovery<Client, Network, Block> where
Block: BlockT + Unpin + 'static,
Network: NetworkProvider,
Client: ProvideRuntimeApi<Block> + Send + Sync + 'static + HeaderBackend<Block>,
<Client as ProvideRuntimeApi<Block>>::Api: AuthorityDiscoveryApi<Block, Error = Error>,
Self: Future<Output = ()>,

pub fn new(
client: Arc<Client>,
network: Arc<Network>,
sentry_nodes: Vec<String>,
key_store: BareCryptoStorePtr,
dht_event_rx: Pin<Box<dyn Stream<Item = DhtEvent> + Send>>
) -> Self

Return a new authority discovery.

Note: When specifying sentry_nodes this module will not advertise the public addresses of the node itself but only the public addresses of its sentry nodes.